eulertest, P162, P267: Fixed code logic for Python 2. eulerlib: Tweaked prime_generator() to be faster and avoid calling xrange() with an argument over 2^31 due to Python 2's limitations (e.g. for P381).

+++ python/eulertest.py
@@ -17,8 +17,8 @@ def main():
 		actualans = module.compute()
 		elapsedtime = time.time() - starttime
 		totaltime += elapsedtime
-		print("Problem {:03d}: {:7d} ms{}".format(prob, round(elapsedtime * 1000), "" if actualans == expectans else "    *** FAIL ***"))
-	print("Total computation time: {} ms".format(round(totaltime * 1000)))
+		print("Problem {:03d}: {:7d} ms{}".format(prob, int(round(elapsedtime * 1000)), "" if actualans == expectans else "    *** FAIL ***"))
+	print("Total computation time: {} ms".format(int(round(totaltime * 1000))))
 
 
 ANSWERS = {
